services/backlinks/requirements.txt
psycopg2-binary
python-dotenv
apscheduler
numpy
